Extract CaptureSnapshotView
