+++ Snap-O/CaptureWindow/CaptureWindow.swift
@@ -1,5 +1,3 @@
-import AppKit
-import AVFoundation
 import SwiftUI
 
 struct CaptureWindow: View {
@@ -9,16 +7,11 @@ struct CaptureWindow: View {
     ZStack {
       Color.black.ignoresSafeArea()
 
-      let transition = transition(for: controller.transitionDirection)
-
-      if let capture = controller.currentCapture {
-        CaptureMediaView(
-          controller: controller,
-          capture: capture
+      if controller.currentCapture != nil {
+        CaptureSnapshotView(
+          controller: controller.snapshotController,
+          livePreviewHost: controller
         )
-        .id(controller.currentCaptureViewID)
-        .zIndex(1)
-        .transition(transition)
       } else if controller.isDeviceListInitialized {
         IdleOverlayView(
           hasDevices: controller.hasDevices,
@@ -63,38 +56,5 @@ struct CaptureWindow: View {
         }
       }
     }
-    .overlay(alignment: .top) {
-      if controller.mediaList.count > 1 {
-        let captures = controller.overlayMediaList.isEmpty ? controller.mediaList : controller.overlayMediaList
-        CapturePreviewStrip(
-          captures: captures,
-          selectedID: controller.selectedMediaID
-        ) { controller.selectMedia(id: $0) }
-          .opacity(controller.shouldShowPreviewHint ? 1 : 0)
-          .offset(y: controller.shouldShowPreviewHint ? 0 : -20)
-          .padding(.top, 12)
-          .allowsHitTesting(controller.shouldShowPreviewHint)
-          .onHover { controller.setPreviewHintHovering($0) }
-          .animation(.easeInOut(duration: 0.35), value: controller.shouldShowPreviewHint)
-      }
-    }
-    .animation(.snappy(duration: 0.25), value: controller.currentCaptureViewID)
-  }
-}
-
-extension CaptureWindow {
-  private func transition(for direction: DeviceTransitionDirection) -> AnyTransition {
-    switch direction {
-    case .previous: xTransition(insertion: .leading, removal: .trailing)
-    case .next: xTransition(insertion: .trailing, removal: .leading)
-    case .neutral: .opacity
-    }
-  }
-
-  private func xTransition(insertion: Edge, removal: Edge) -> AnyTransition {
-    .asymmetric(
-      insertion: .move(edge: insertion).combined(with: .opacity),
-      removal: .move(edge: removal).combined(with: .opacity)
-    )
   }
 }
